Agrego ejemplo-para-tdd.txt y cambios en tdd y solid

+++ tdd/src/main/java/org/ehmsoft/Calculadora.java
@@ -1,12 +1,13 @@
 package org.ehmsoft;
 
 public class Calculadora {
-    public static int sumar(int a, int b) {
+    public static double sumar(int a, int b) {
         return a + b;
     }
 
-    public static int restar(int a, int b) {
+    public static double restar(int a, int b) {
         return a - b;
     }
+
 }
